Add game formats and associate Rooms / Board with them
