+++ src/client/components/DeckBuilderFilters/DeckBuilderFilters.tsx
@@ -204,7 +204,7 @@ const RaritiesFilter: React.FC = () => {
                                 points="7,1 13,7 7,13 1,7"
                                 fill={COLORS_FOR_RARITY[rarity]}
                                 stroke="white"
-                                stroke-width="1"
+                                strokeWidth="1"
                             />
                         </svg>
                     </CastingCostFrame>
