+++ src/client/components/RoomSquare/RoomSquare.spec.tsx
@@ -2,19 +2,20 @@ import { fireEvent, render, screen } from '@testing-library/react';
 import React from 'react';
 
 import { RoomSquare } from './RoomSquare';
+import { Format } from '@/types/games';
+import { DetailedRoom } from '@/types';
+
+const detailedRoom: DetailedRoom = {
+    roomName: 'Room 6',
+    players: ['Kimmy', 'Jimmy', 'Timmy'],
+    spectators: [],
+    avatarsForPlayers: {},
+    format: Format.STANDARD,
+};
 
 describe('Room Square', () => {
     it('renders a room name + players', () => {
-        render(
-            <RoomSquare
-                detailedRoom={{
-                    roomName: 'Room 6',
-                    players: ['Kimmy', 'Jimmy', 'Timmy'],
-                    spectators: [],
-                    avatarsForPlayers: {},
-                }}
-            />
-        );
+        render(<RoomSquare detailedRoom={detailedRoom} />);
         expect(screen.getByText('Room 6')).toBeInTheDocument();
         expect(screen.getByText('Kimmy')).toBeInTheDocument();
         expect(screen.getByText('Jimmy')).toBeInTheDocument();
@@ -22,16 +23,7 @@ describe('Room Square', () => {
     });
 
     it('truncates "public-"', () => {
-        render(
-            <RoomSquare
-                detailedRoom={{
-                    roomName: 'public-Room 6',
-                    players: ['Kimmy', 'Jimmy', 'Timmy'],
-                    spectators: [],
-                    avatarsForPlayers: {},
-                }}
-            />
-        );
+        render(<RoomSquare detailedRoom={detailedRoom} />);
         expect(screen.getByText('Room 6')).toBeInTheDocument();
     });
 
@@ -40,13 +32,7 @@ describe('Room Square', () => {
             const mockRejoinGame = jest.fn();
             render(
                 <RoomSquare
-                    detailedRoom={{
-                        hasStartedGame: true,
-                        roomName: 'public-Room 6',
-                        players: ['Kimmy', 'Jimmy', 'Timmy'],
-                        spectators: [],
-                        avatarsForPlayers: {},
-                    }}
+                    detailedRoom={{ ...detailedRoom, hasStartedGame: true }}
                     rejoinRoom={mockRejoinGame}
                     hasJoined
                 />
@@ -59,13 +45,7 @@ describe('Room Square', () => {
             const mockRejoinGame = jest.fn();
             render(
                 <RoomSquare
-                    detailedRoom={{
-                        hasStartedGame: true,
-                        roomName: 'public-Room 6',
-                        players: ['Kimmy', 'Jimmy', 'Timmy'],
-                        spectators: [],
-                        avatarsForPlayers: {},
-                    }}
+                    detailedRoom={detailedRoom}
                     rejoinRoom={mockRejoinGame}
                 />
             );
@@ -76,12 +56,7 @@ describe('Room Square', () => {
             const mockRejoinGame = jest.fn();
             render(
                 <RoomSquare
-                    detailedRoom={{
-                        roomName: 'public-Room 6',
-                        players: ['Kimmy', 'Jimmy', 'Timmy'],
-                        spectators: [],
-                        avatarsForPlayers: {},
-                    }}
+                    detailedRoom={detailedRoom}
                     rejoinRoom={mockRejoinGame}
                     hasJoined
                 />
@@ -95,12 +70,7 @@ describe('Room Square', () => {
             const mockSpectateRoom = jest.fn();
             render(
                 <RoomSquare
-                    detailedRoom={{
-                        roomName: 'public-Room 6',
-                        players: ['Kimmy', 'Jimmy', 'Timmy'],
-                        spectators: [],
-                        avatarsForPlayers: {},
-                    }}
+                    detailedRoom={detailedRoom}
                     spectateRoom={mockSpectateRoom}
                 />
             );
@@ -109,17 +79,7 @@ describe('Room Square', () => {
         });
 
         it('hides the spectate button if the game is already spectated', () => {
-            render(
-                <RoomSquare
-                    detailedRoom={{
-                        roomName: 'public-Room 6',
-                        players: ['Kimmy', 'Jimmy', 'Timmy'],
-                        spectators: [],
-                        avatarsForPlayers: {},
-                    }}
-                    isSpectacting
-                />
-            );
+            render(<RoomSquare detailedRoom={detailedRoom} isSpectacting />);
             expect(screen.queryByText('Spectate')).not.toBeInTheDocument();
         });
 
@@ -127,11 +87,8 @@ describe('Room Square', () => {
             render(
                 <RoomSquare
                     detailedRoom={{
-                        roomName: 'public-Room 6',
-                        players: ['Kimmy', 'Jimmy', 'Timmy'],
-                        spectators: [],
+                        ...detailedRoom,
                         hasStartedGame: true,
-                        avatarsForPlayers: {},
                     }}
                 />
             );
@@ -144,12 +101,7 @@ describe('Room Square', () => {
             const mockJoinRoom = jest.fn();
             render(
                 <RoomSquare
-                    detailedRoom={{
-                        roomName: 'public-Room 6',
-                        players: ['Kimmy', 'Jimmy', 'Timmy'],
-                        spectators: [],
-                        avatarsForPlayers: {},
-                    }}
+                    detailedRoom={detailedRoom}
                     joinRoom={mockJoinRoom}
                 />
             );
@@ -166,6 +118,7 @@ describe('Room Square', () => {
                         players: ['Kimmy', 'Jimmy', 'Timmy'],
                         spectators: [],
                         avatarsForPlayers: {},
+                        format: Format.STANDARD,
                     }}
                     hasJoined
                     joinRoom={mockJoinRoom}
@@ -183,6 +136,7 @@ describe('Room Square', () => {
                         players: ['Kimmy', 'Jimmy', 'Timmy', 'Player 4'],
                         spectators: [],
                         avatarsForPlayers: {},
+                        format: Format.STANDARD,
                     }}
                     joinRoom={mockJoinRoom}
                 />
@@ -194,13 +148,7 @@ describe('Room Square', () => {
     it('renders Started if a game has started', () => {
         render(
             <RoomSquare
-                detailedRoom={{
-                    roomName: 'Room 6',
-                    players: ['Kimmy', 'Jimmy', 'Timmy'],
-                    spectators: [],
-                    hasStartedGame: true,
-                    avatarsForPlayers: {},
-                }}
+                detailedRoom={{ ...detailedRoom, hasStartedGame: true }}
             />
         );
         expect(screen.getByText('Started')).toBeInTheDocument();
@@ -216,6 +164,7 @@ describe('Room Square', () => {
                     spectators: [],
                     hasStartedGame: false,
                     avatarsForPlayers: {},
+                    format: Format.STANDARD,
                 }}
                 hasJoined
                 onStartGameClicked={mockStartGame}
@@ -229,13 +178,7 @@ describe('Room Square', () => {
         const mockStartGame = jest.fn();
         render(
             <RoomSquare
-                detailedRoom={{
-                    roomName: 'Room 6',
-                    players: ['Kimmy', 'Jimmy'],
-                    spectators: [],
-                    hasStartedGame: true,
-                    avatarsForPlayers: {},
-                }}
+                detailedRoom={{ ...detailedRoom, hasStartedGame: true }}
                 hasJoined
                 onStartGameClicked={mockStartGame}
             />
@@ -249,11 +192,8 @@ describe('Room Square', () => {
         render(
             <RoomSquare
                 detailedRoom={{
-                    roomName: 'Room 6',
-                    players: ['Kimmy', 'Jimmy', 'Timmy'],
-                    spectators: [],
+                    ...detailedRoom,
                     hasStartedGame: false,
-                    avatarsForPlayers: {},
                 }}
                 hasJoined
                 onStartGameClicked={mockStartGame}
